Updated the MSVC project files and ensured that everything works (especially the most recent fix to Screen).

+++ app/example/disasteroids/dis_master.cpp
@@ -80,7 +80,7 @@ void Master::Run ()
     ResourceCache resource_cache;
 
     // seed the random number generator
-    srand(time(NULL));
+    srand(static_cast<Uint32>(time(NULL)));
 
     // start at the title screen
     ActivateTitleScreen();
